server-requirements.txt
Django==1.5.3
distribute==0.6.28
djangorestframework==2.3.8
wsgiref==0.1.2
